Resolved initialization and function scope issues.
